requirements.txt
Django
djangorestframework
psycopg2
strawberry-graphql-django
